+++ CHANGELOG.md
@@ -204,6 +204,7 @@ IMPROVEMENTS:
 * Introduce penalty based scheduler
 * Drop unused redundant support for Google PubSub
 * Added a stripped down CPU single node cluster for AWS deployment example in examples/aws/cpu, complete with helloworld studioml python code.
+* Draft : Message payload encryption supported across insecure transports
 
 FIXES: